# docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  models.Student:
    properties:
      birth_date:
        type: string
      book_id:
        type: integer
      id:
        type: integer
      middle_name:
        type: string
      name:
        type: string
      student_group:
        type: string
      surname:
        type: string
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
      successful:
        type: boolean
    type: object
  models.UpdateEventRequest:
    properties:
      event_id:
        type: integer
      new_data:
        properties:
          description:
            type: string
          event_date:
            type: string
          event_type_code:
            type: integer
          icon_url:
            type: string
          points:
            type: integer
          title:
            type: string
        type: object
    type: object
...
  models.User:
    properties:
      birth_date:
        type: string
      book_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      middle_name:
        type: string
      name:
        type: string
      password:
        items:
          type: integer
        type: array
      role_level:
        type: integer
      student_group:
        type: string
      surname:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /admin/students:
    get:
      description: Возвращает всех студентов из таблицы students.
      parameters:
      - default: Bearer
        description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список студентов
          schema:
            items:
              $ref: '#/definitions/models.Student'
            type: array
        "500":
          description: Ошибка при запросе или чтении данных
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка студентов
      tags:
      - admin
  /admin/update_event:
    patch:
      consumes:
      - application/json
      description: |-
        Производит частичное обновление данных события по его ID.
        Обновляются только те поля, которые переданы в теле запроса.
      parameters:
      - description: 'Bearer токен авторизации. Формат: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID события и новые значения полей
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Событие успешно обновлено
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Некорректный JSON или ошибка валидации
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ — неверный или отсутствующий токен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Ошибка сервера при попытке обновления записи
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Обновить выбранные поля события
      tags:
      - admin
...
  /admin/users:
    get:
      description: Возвращает всех зарегистрированных пользователей из таблицы users.
      parameters:
      - default: Bearer
        description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Ошибка при запросе или чтении данных
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка пользователей
      tags:
      - admin
...
swagger: "2.0"